mcp-agent-services/crawl4ai-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
crawl4ai>=0.2.0
playwright>=1.40.0
feedparser>=6.0.10
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
